+++ src/project/template/uni_modules/mp-html/components/mp-html/parser.js
@@ -10,6 +10,11 @@ const config = {
   // 块级标签（转为 div，其他的非信任标签转为 span）
   blockTags: makeMap('address,article,aside,body,caption,center,cite,footer,header,html,nav,pre,section'),
 
+  // #ifdef (MP-WEIXIN || MP-QQ || APP-PLUS || MP-360) && VUE3
+  // 行内标签
+  inlineTags: makeMap('abbr,b,big,code,del,em,i,ins,label,q,small,span,strong,sub,sup'),
+  // #endif
+
   // 要移除的标签
   ignoreTags: makeMap('area,base,canvas,embed,frame,head,iframe,input,link,map,meta,param,rp,script,source,style,textarea,title,track,wbr'),
 
@@ -34,7 +39,11 @@ const config = {
     ldquo: '“',
     rdquo: '”',
     bull: '•',
-    hellip: '…'
+    hellip: '…',
+    larr: '←',
+    uarr: '↑',
+    rarr: '→',
+    darr: '↓'
   },
 
   // 默认的标签样式
@@ -496,15 +505,11 @@ Parser.prototype.onOpenTag = function (selfClose) {
         styleObj.height = undefined
       }
       // 记录是否设置了宽高
-      if (styleObj.width) {
-        if (styleObj.width.includes('auto')) {
-          styleObj.width = ''
-        } else {
-          node.w = 'T'
-          if (styleObj.height && !styleObj.height.includes('auto')) {
-            node.h = 'T'
-          }
-        }
+      if (!isNaN(parseInt(styleObj.width))) {
+        node.w = 'T'
+      }
+      if (!isNaN(parseInt(styleObj.height)) && (!styleObj.height.includes('%') || (parent && (parent.attrs.style || '').includes('height')))) {
+        node.h = 'T'
       }
     } else if (node.name === 'svg') {
       siblings.push(node)
@@ -695,7 +700,7 @@ Parser.prototype.popNode = function () {
         } else if (size > 7) {
           size = 7
         }
-        styleObj['font-size'] = ['xx-small', 'x-small', 'small', 'medium', 'large', 'x-large', 'xx-large'][size - 1]
+        styleObj['font-size'] = ['x-small', 'small', 'medium', 'large', 'x-large', 'xx-large', 'xxx-large'][size - 1]
       }
       attrs.size = undefined
     }
@@ -728,7 +733,11 @@ Parser.prototype.popNode = function () {
     // #endif
   ) {
     this.expose()
-  } /* #ifdef APP-PLUS */ else if (node.name === 'video') {
+  } else if (node.name === 'video') {
+    if ((styleObj.height || '').includes('auto')) {
+      styleObj.height = undefined
+    }
+    /* #ifdef APP-PLUS */
     let str = '<video style="width:100%;height:100%"'
     for (const item in attrs) {
       if (attrs[item]) {
@@ -736,7 +745,7 @@ Parser.prototype.popNode = function () {
       }
     }
     if (this.options.pauseVideo) {
-      str += ' onplay="for(var e=document.getElementsByTagName(\'video\'),t=0;t<e.length;t++)e[t]!=this&&e[t].pause()"'
+      str += ' onplay="this.dispatchEvent(new CustomEvent(\'vplay\',{bubbles:!0}));for(var e=document.getElementsByTagName(\'video\'),t=0;t<e.length;t++)e[t]!=this&&e[t].pause()"'
     }
     str += '>'
     for (let i = 0; i < node.src.length; i++) {
@@ -744,7 +753,8 @@ Parser.prototype.popNode = function () {
     }
     str += '</video>'
     node.html = str
-  } /* #endif */ else if ((node.name === 'ul' || node.name === 'ol') && node.c) {
+    /* #endif */
+  } else if ((node.name === 'ul' || node.name === 'ol') && node.c) {
     // 列表处理
     const types = {
       a: 'lower-alpha',
@@ -807,7 +817,7 @@ Parser.prototype.popNode = function () {
 
       for (let row = 1; row <= trList.length; row++) {
         let col = 1
-        for (let j = 0; j < trList[row - 1].children.length; j++, col++) {
+        for (let j = 0; j < trList[row - 1].children.length; j++) {
           const td = trList[row - 1].children[j]
           if (td.name === 'td' || td.name === 'th') {
             // 这个格子被上面的单元格占用，则列号++
@@ -853,6 +863,7 @@ Parser.prototype.popNode = function () {
               td.attrs.style = style
             }
             cells.push(td)
+            col++
           }
         }
         if (row === 1) {
@@ -932,7 +943,13 @@ Parser.prototype.popNode = function () {
   } else if (node.c) {
     node.c = 2
     for (let i = node.children.length; i--;) {
-      if (!node.children[i].c || node.children[i].name === 'table') {
+      const child = node.children[i]
+      // #ifdef (MP-WEIXIN || MP-QQ || APP-PLUS || MP-360) && VUE3
+      if (child.name && (config.inlineTags[child.name] || (child.attrs.style || '').includes('inline'))) {
+        child.c = 1
+      }
+      // #endif
+      if (!child.c || child.name === 'table') {
         node.c = 1
       }
     }
@@ -948,7 +965,7 @@ Parser.prototype.popNode = function () {
     }
   }
   // flex 布局时部分样式需要提取到 rich-text 外层
-  const flex = parent && (parent.attrs.style || '').includes('flex')
+  const flex = parent && ((parent.attrs.style || '').includes('flex') || (parent.attrs.style || '').includes('grid'))
     // #ifdef MP-WEIXIN
     // 检查基础库版本 virtualHost 是否可用
     && !(node.c && wx.getNFCAdapter) // eslint-disable-line
@@ -959,13 +976,31 @@ Parser.prototype.popNode = function () {
   if (flex) {
     node.f = ';max-width:100%'
   }
+
+  // 优化长内容加载速度
+  if (children.length >= 50 && node.c && !(styleObj.display || '').includes('flex')) {
+    let i = children.length - 1
+    for (let j = i; j >= -1; j--) {
+      // 合并多个块级标签
+      if (j === -1 || children[j].c || !children[j].name || (children[j].name !== 'div' && children[j].name !== 'p' && children[j].name[0] !== 'h') || (children[j].attrs.style || '').includes('inline')) {
+        if (i - j >= 5) {
+          children.splice(j + 1, i - j, {
+            name: 'div',
+            attrs: {},
+            children: node.children.slice(j + 1, i + 1)
+          })
+        }
+        i = j - 1
+      }
+    }
+  }
   // #endif
 
   for (const key in styleObj) {
     if (styleObj[key]) {
       const val = `;${key}:${styleObj[key].replace(' !important', '')}`
       /* #ifndef APP-PLUS-NVUE */
-      if (flex && ((key.includes('flex') && key !== 'flex-direction') || key === 'align-self' || styleObj[key][0] === '-' || (key === 'width' && val.includes('%')))) {
+      if (flex && ((key.includes('flex') && key !== 'flex-direction') || key === 'align-self' || key.includes('grid') || styleObj[key][0] === '-' || (key.includes('width') && val.includes('%')))) {
         node.f += val
         if (key === 'width') {
           attrs.style += ';width:100%'
@@ -976,6 +1011,11 @@ Parser.prototype.popNode = function () {
     }
   }
   attrs.style = attrs.style.substr(1) || undefined
+  // #ifdef (MP-WEIXIN || MP-QQ) && VUE3
+  if (!attrs.style) {
+    delete attrs.style
+  }
+  // #endif
 }
 
 /**
@@ -1005,6 +1045,9 @@ Parser.prototype.onText = function (text) {
   }
   const node = Object.create(null)
   node.type = 'text'
+  // #ifdef (MP-BAIDU || MP-ALIPAY || MP-TOUTIAO) && VUE3
+  node.attrs = {}
+  // #endif
   node.text = decodeEntity(text)
   if (this.hook(node)) {
     // #ifdef MP-WEIXIN
@@ -1220,4 +1263,4 @@ Lexer.prototype.endTag = function () {
   }
 }
 
-module.exports = Parser
+export default Parser